requirements.txt
Django==5.1.3
djangorestframework==3.15.2
django-filter==24.3
python-dotenv==1.0.1
psycopg2==2.9.10
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
pillow==11.0.0
django-debug-toolbar==4.4.6
django-filter==24.3
huggingface-hub==0.27.0
openai==0.28.0
requests==2.32.3
selenium==4.27.1
undetected-chromedriver==3.5.5
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==14.1